+++ sdk/core/azure-core/src/http/policy.cpp
@@ -7,16 +7,6 @@
 using Azure::Core::Context;
 using namespace Azure::Core::Http;
 
-#ifndef _MSC_VER
-// Non-MSVC compilers do require allocation of statics, even if they are const constexpr.
-// MSVC, on the other hand, has problem if you "redefine" static constexprs.
-Azure::Core::Logging::LogClassification const Azure::Core::Http::LogClassification::Request;
-Azure::Core::Logging::LogClassification const Azure::Core::Http::LogClassification::Response;
-Azure::Core::Logging::LogClassification const Azure::Core::Http::LogClassification::Retry;
-Azure::Core::Logging::LogClassification const
-    Azure::Core::Http::LogClassification::HttpTransportAdapter;
-#endif
-
 // The NextHttpPolicy can't be created from a nullptr because it is a reference. So we don't need to
 // check if m_policies is nullptr.
 std::unique_ptr<RawResponse> NextHttpPolicy::Send(Context const& ctx, Request& req)