+++ sdk/storage/azure-storage-files-shares/test/share_file_client_test.cpp
@@ -7,6 +7,7 @@
 #include <chrono>
 #include <future>
 
+#include <azure/core/cryptography/hash.hpp>
 #include <azure/storage/common/crypt.hpp>
 #include <azure/storage/common/file_io.hpp>
 #include <azure/storage/common/storage_common.hpp>
@@ -261,8 +262,8 @@ namespace Azure { namespace Storage { namespace Test {
   {
     auto result = m_fileClient->ListHandlesSinglePage();
     EXPECT_TRUE(result->Handles.empty());
-    EXPECT_TRUE(result->ContinuationToken.empty());
-    EXPECT_NO_THROW(m_fileClient->ForceCloseAllHandles());
+    EXPECT_FALSE(result->ContinuationToken.HasValue());
+    EXPECT_NO_THROW(m_fileClient->ForceCloseAllHandlesSinglePage());
   }
 
   TEST_F(FileShareFileClientTest, LeaseRelated)
@@ -600,6 +601,12 @@ namespace Azure { namespace Storage { namespace Test {
     }
   }
 
+  static std::vector<uint8_t> Hash(const std::string& data)
+  {
+    Azure::Core::Cryptography::Md5Hash instance;
+    return instance.Final(reinterpret_cast<const uint8_t*>(data.data()), data.size());
+  }
+
   TEST_F(FileShareFileClientTest, RangeUploadDownload)
   {
     auto rangeSize = 1 * 1024 * 1024;
@@ -640,8 +647,9 @@ namespace Azure { namespace Storage { namespace Test {
     {
       // MD5 works.
       memBodyStream.Rewind();
-      auto md5 = Md5::Hash(rangeContent.data(), rangeContent.size());
-      auto invalidMd5 = Md5::Hash(std::string("This is garbage."));
+      Azure::Core::Cryptography::Md5Hash instance;
+      auto md5 = instance.Final(rangeContent.data(), rangeContent.size());
+      auto invalidMd5 = Hash(std::string("This is garbage."));
       auto fileClient
           = m_shareClient->GetRootDirectoryClient().GetFileClient(LowercaseRandomString(10));
       Files::Shares::UploadShareFileRangeOptions uploadOptions;